+++ resolvers/webpack/test/externals.js
@@ -9,6 +9,13 @@ const webpack = require('../index');
 const file = path.join(__dirname, 'files', 'dummy.js');
 
 describe('externals', function () {
+  const settingsWebpack5 = {
+    config: require(path.join(__dirname, './files/webpack.config.webpack5.js')),
+  };
+  const settingsWebpack5Async = {
+    config: require(path.join(__dirname, './files/webpack.config.webpack5.async-externals.js')),
+  };
+
   it('works on just a string', function () {
     const resolved = webpack.resolve('bootstrap', file);
     expect(resolved).to.have.property('found', true);
@@ -31,5 +38,27 @@ describe('externals', function () {
     const resolved = webpack.resolve('fs', file);
     expect(resolved).to.have.property('found', true);
     expect(resolved).to.have.property('path', null);
+  });
+
+  it('works on a function (synchronous) for webpack 5', function () {
+    const resolved = webpack.resolve('underscore', file, settingsWebpack5);
+    expect(resolved).to.have.property('found', true);
+    expect(resolved).to.have.property('path', null);
+  });
+
+  it('works on a function (synchronous) which uses getResolve for webpack 5', function () {
+    const resolved = webpack.resolve('graphql', file, settingsWebpack5);
+    expect(resolved).to.have.property('found', true);
+    expect(resolved).to.have.property('path', null);
+  });
+
+  it('prevents using an asynchronous function for webpack 5', function () {
+    const resolved = webpack.resolve('underscore', file, settingsWebpack5Async);
+    expect(resolved).to.have.property('found', false);
+  });
+
+  it('prevents using a function which uses Promise returned by getResolve for webpack 5', function () {
+    const resolved = webpack.resolve('graphql', file, settingsWebpack5Async);
+    expect(resolved).to.have.property('found', false);
   });
 });